Reworked the cron part to use the WHMCS task setup.
Cronjobs can also be enabled/disabled by the admin user

Fixes #34
