+++ modules/addons/realtimeregister_ssl/hooks.php
@@ -766,3 +766,20 @@ add_hook('DailyCronJob', 10, function () {
     curl_exec($ch);
     curl_close($ch);
 });
+
+
+// Setup tasks, these can be disabled via the admin
+global $CONFIG;
+
+require_once __DIR__ . DS . 'Loader.php';
+new Loader();
+
+\AddonModule\RealtimeRegisterSsl\cron\AutomaticSynchronisation::register();
+\AddonModule\RealtimeRegisterSsl\cron\ProcessingOrders::register();
+\AddonModule\RealtimeRegisterSsl\cron\DailyStatusUpdater::register();
+\AddonModule\RealtimeRegisterSsl\cron\CertificateStatisticsLoader::register();
+\AddonModule\RealtimeRegisterSsl\cron\Notifier::register();
+\AddonModule\RealtimeRegisterSsl\cron\CertificateSender::register();
+\AddonModule\RealtimeRegisterSsl\cron\PriceUpdater::register();
+\AddonModule\RealtimeRegisterSsl\cron\CertificateDetailsUpdater::register();
+\AddonModule\RealtimeRegisterSsl\cron\InstallCertificates::register();
