+++ modules/addons/realtimeregister_ssl/templates/admin/pages/crons/crons.tpl
@@ -1,0 +1,10 @@
+<div class="panel panel-default">
+    <div class="panel-heading">{$ADDONLANG->T('header')}</div>
+    <div class="panel-body">{$ADDONLANG->T('explanation')}</div>
+</div>
+<div class="panel panel-default">
+    <div class="panel-heading">{$ADDONLANG->T('header')}</div>
+    <div class="panel-body">
+        {$form}
+    </div>
+</div>
